Fix a buffer overflow in FormatPath.
FormatPath assumes that it can read and write beyond the end of the
passed in string. However, the buffers passed to FormatPath are not
sized appropriately for this to be valid.

Change FormatPath to return a newly allocated buffer of the appropriate
size and to only read within the length of the string.

The call to FormatPath in ReadPathsFromFile is also moved to
SetPathFile.